requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.1
geoalchemy2==0.14.1
streamlit==1.28.1
folium==0.14.0
plotly==5.18.0
requests==2.31.0
reportlab==4.0.7
